+++ test/__fixtures/utils/server-helper.js
@@ -1,0 +1,9 @@
+const happn = require('../../../lib/index');
+module.exports = class ServerHelper {
+    static create () {
+        return new ServerHelper();
+    }
+    async createServer(config) {
+        return await happn.service.create(config);
+    }
+}